server started

+++ ConsoleApp/Program.cs
@@ -1,8 +1,12 @@
 ﻿using System;
+using System.IO;
 using ClassLibrary;
 using System.Collections.Generic;
 using Microsoft.Data.Sqlite;
 using Terminal.Gui;
+using System.Net;
+using System.Net.Sockets;
+using System.Text;
 
 namespace ConsoleApp
 {
@@ -21,8 +25,28 @@ namespace ConsoleApp
         static ListView answersListView;
         static Label page;
         static Label totalPages;
+        static IPAddress ipAddress = IPAddress.Loopback;   
+        static Socket sender = new Socket(ipAddress.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
+        static IPEndPoint remoteEP = new IPEndPoint(ipAddress, 3000);
         static void Main(string[] args)
         {
+            try
+            {
+                sender.Connect(remoteEP);
+
+                Console.WriteLine("Socket connected to {0}", sender.RemoteEndPoint.ToString());
+
+                ProcessWhileConnected(sender);
+
+                // Release the socket.  
+                //sender.Shutdown(SocketShutdown.Both);
+                sender.Close();
+
+            }
+            catch (Exception)
+            {
+                //Console.WriteLine("Unexpected exception : {0}", e.ToString());
+            }
             Application.Init();
             Window win = new Window("Log In")
             {
@@ -46,23 +70,70 @@ namespace ConsoleApp
             win.Add(pLabel, inputPass);
             Application.Run();
         }
+        static void ProcessWhileConnected(Socket sender)
+        {
+            // Encode the data string into a byte array. 
+            //while (true)
+            //{
+                try
+                {
+                    //Console.WriteLine("Send your message ");
+                    string input = "Hello from App";//Console.ReadLine();
+                    byte[] msg = Encoding.ASCII.GetBytes(input);
+
+                    // Send the data through the socket.  
+                    Console.WriteLine("Sending: " + input);
+                    if (input == "Exit")
+                    {
+                        //sender.Close();
+                        //break;
+                    }
+                    int bytesSent = sender.Send(msg);
+
+                    byte[] bytes = new byte[1024];  // buffer
+                    int bytesRec = sender.Receive(bytes);
+                    Console.WriteLine("Server send: " + Encoding.ASCII.GetString(bytes, 0, bytesRec));
+                }
+                catch (Exception e)
+                {
+                    Console.WriteLine("Unexpected exception : {0}", e.ToString());
+                }
+                //byte[] bytes = new byte[1024];
+                //int bytesRec = sender.Receive(bytes);
+                //string message = Encoding.ASCII.GetString(bytes, 0, bytesRec);
+                //Console.WriteLine("Get:\n", message);
+            //}
+        }
         static void OnButtonRegClicked()
         {
-            RegistrationDialog dialog = new RegistrationDialog();
-            Application.Run(dialog);
-            if (!dialog.canceled)
+            if(!File.Exists(dbPath))
             {
-                User user = dialog.GetUser();
-                if (userRepository.FindLogin(user.login) == null) 
+                MessageBox.ErrorQuery("Error", "DataBase not found.", "Ok");
+                Application.RequestStop();
+            }
+            else
+            {
+                RegistrationDialog dialog = new RegistrationDialog();
+                Application.Run(dialog);
+                if (!dialog.canceled)
                 {
-                    Autentification.Register(userRepository, user.name, user.login, user.isModerator, user.password);
+                    User user = dialog.GetUser();
+                    if (userRepository.FindLogin(user.login) == null) 
+                    {
+                        Autentification.Register(userRepository, user.name, user.login, user.isModerator, user.password);
+                    }
+                    else 
+                        MessageBox.Query("Error", "You can't use this login. It is already taken.", "Ok");
                 }
-                else 
-                    MessageBox.Query("Error", "You can't use this login. It is already taken.", "Ok");
             }
         }
         static void OnButtonVerifyClicked()
         {
+            if(!File.Exists(dbPath))
+            {
+                MessageBox.ErrorQuery("Error", "DataBase not found.", "Ok");
+                Application.RequestStop();
+            }
             current = Autentification.Verify(userRepository, inputLogin.Text.ToString(), inputPass.Text.ToString());
             if (current != null)
                 OnButtonUsersClicked();
@@ -77,6 +148,7 @@ namespace ConsoleApp
                 new MenuBarItem ("File", new MenuItem [] {
                     new MenuItem ("Import", "", OnImportClicked),
                     new MenuItem ("Export", "", OnExportClicked),
+                    new MenuItem ("Get stats", "", OnGetImage),
                     new MenuItem ("Exit", "", OnExit),
                 }),
                 new MenuBarItem ("Help", new MenuItem [] {
@@ -91,12 +163,14 @@ namespace ConsoleApp
                 Height = Dim.Fill() - 1
             };
             top.Add(menu, win);
-            Label userLabel = new Label(80, 1, $"You: {current.name}({current.login})" );
+            Label userLabel = new Label(70, 1, $"You: {current.name}({current.login})" );
+            Button btnex = new Button(100, 1, "Change user");
+            btnex.Clicked += OnChangeUser;
             Button btnq = new Button(1, 1, "Go to questions");
             btnq.Clicked += OnButtonQuestionsClicked;
             Button btna = new Button(20, 1, "Go to answers");
             btna.Clicked += OnButtonAnswersClicked;
-            win.Add(btnq, btna, userLabel);
+            win.Add(btnq, btna, btnex, userLabel);
             Rect frame = new Rect(4, 8, top.Frame.Width, 200);
             if (userRepository.GetTotalPages() < 1) 
             {
@@ -152,6 +226,11 @@ namespace ConsoleApp
             Application.Run(dialog);
             if (dialog.deleted)
             {
+                if (user.id != current.id && current.isModerator == "no")
+                {
+                    MessageBox.Query("Error", "You have no access to this function", "Ok");
+                    return;
+                }
                 int pageNumber = Int32.Parse(Convert.ToString(page.Text));
                 userRepository.DeleteById(user.id);
                 int total = userRepository.GetTotalPages();
@@ -171,10 +250,20 @@ namespace ConsoleApp
             }
             if (dialog.updated)
             {
+                if (user.id != current.id && current.isModerator == "no")
+                {
+                    MessageBox.Query("Error", "You have no access to this function", "Ok");
+                    return;
+                }
                 int userId = user.id;
                 int pageNumber = userId/10;
                 if(userId % 10 != 0) pageNumber++;
                 user = dialog.GetUser();
+                if(user.name == "" || user.login == "" || user.password == "") 
+                {
+                    MessageBox.Query("Error", "Wrong input", "Ok");
+                    return;
+                }
                 userRepository.Update(userId, user);
                 usersListView.SetSource(userRepository.GetPage(pageNumber));
             }
@@ -186,7 +275,12 @@ namespace ConsoleApp
             if(!dialog.canceled)
             {
                 User user = dialog.GetUser();
-                userRepository.Insert(user);
+                if(user.name == "" || user.login == "" || user.password == "") 
+                {
+                    MessageBox.Query("Error", "Wrong input", "Ok");
+                    return;
+                }
+                Autentification.Register(userRepository, user.name, user.login, user.isModerator, user.password);
                 usersListView.SetSource(userRepository.GetPage(userRepository.GetTotalPages()));
                 page.Text = Convert.ToString(userRepository.GetTotalPages());
                 totalPages.Text = Convert.ToString(userRepository.GetTotalPages());
@@ -201,6 +295,7 @@ namespace ConsoleApp
                 new MenuBarItem ("File", new MenuItem [] {
                     new MenuItem ("Import", "", OnImportClicked),
                     new MenuItem ("Export", "", OnExportClicked),
+                    new MenuItem ("Get stats", "", OnGetImage),
                     new MenuItem ("Exit", "", OnExit),
                 }),
                 new MenuBarItem ("Help", new MenuItem [] {
@@ -215,12 +310,14 @@ namespace ConsoleApp
                 Height = Dim.Fill() - 1
             };
             top.Add(menu, win);
-            Label userLabel = new Label(80, 1, $"You: {current.name}({current.login})" );
+            Label userLabel = new Label(70, 1, $"You: {current.name}({current.login})" );
+            Button btnex = new Button(100, 1, "Change user");
+            btnex.Clicked += OnChangeUser;
             Button btnu = new Button(1, 1, "Go to users");
             btnu.Clicked += OnButtonUsersClicked;
             Button btna = new Button(20, 1, "Go to answers");
             btna.Clicked += OnButtonAnswersClicked;
-            win.Add(btnu, btna, userLabel);
+            win.Add(btnu, btna, btnex, userLabel);
             Rect frame = new Rect(4, 8, top.Frame.Width, 200);
             if (questionRepository.GetTotalPages() < 1) 
             {
@@ -276,6 +373,11 @@ namespace ConsoleApp
             Application.Run(dialog);
             if (dialog.deleted)
             {
+                if (question.userId != current.id && current.isModerator == "no")
+                {
+                    MessageBox.Query("Error", "You have no access to this function", "Ok");
+                    return;
+                }
                 int pageNumber = Int32.Parse(Convert.ToString(page.Text));
                 questionRepository.DeleteById(question.id);
                 int total = questionRepository.GetTotalPages();
@@ -295,10 +397,20 @@ namespace ConsoleApp
             }
             if (dialog.updated)
             {
+                if (question.userId != current.id && current.isModerator == "no")
+                {
+                    MessageBox.Query("Error", "You have no access to this function", "Ok");
+                    return;
+                }
                 int questionId = question.id;
                 int pageNumber = questionId/10;
                 if(questionId % 10 != 0) pageNumber++;
                 question = dialog.GetQuestion();
+                if(question.title == "" || question.text == "") 
+                {
+                    MessageBox.Query("Error", "Wrong input", "Ok");
+                    return;
+                }
                 questionRepository.Update(questionId, question);
                 questionsListView.SetSource(questionRepository.GetPage(pageNumber));
             }
@@ -310,6 +422,16 @@ namespace ConsoleApp
             if(!dialog.canceled)
             {
                 Question question = dialog.GetQuestion();
+                if(question.title == "" || question.text == "") 
+                {
+                    MessageBox.Query("Error", "Wrong input", "Ok");
+                    return;
+                }
+                if (current.isModerator == "no")
+                {
+                    MessageBox.Query("Warning", "You added a question(urer id was changed)", "Ok");
+                    question.userId = current.id;
+                }
                 questionRepository.Insert(question);
                 questionsListView.SetSource(questionRepository.GetPage(questionRepository.GetTotalPages()));
                 page.Text = Convert.ToString(questionRepository.GetTotalPages());
@@ -325,6 +447,7 @@ namespace ConsoleApp
                 new MenuBarItem ("File", new MenuItem [] {
                     new MenuItem ("Import", "", OnImportClicked),
                     new MenuItem ("Export", "", OnExportClicked),
+                    new MenuItem ("Get stats", "", OnGetImage),
                     new MenuItem ("Exit", "", OnExit),
                 }),
                 new MenuBarItem ("Help", new MenuItem [] {
@@ -339,12 +462,14 @@ namespace ConsoleApp
                 Height = Dim.Fill() - 1
             };
             top.Add(menu, win);
-            Label userLabel = new Label(80, 1, $"You: {current.name}({current.login})" );
+            Label userLabel = new Label(70, 1, $"You: {current.name}({current.login})" );
+            Button btnex = new Button(100, 1, "Change user");
+            btnex.Clicked += OnChangeUser;
             Button btnu = new Button(1, 1, "Go to users");
             btnu.Clicked += OnButtonUsersClicked;
             Button btnq = new Button(20, 1, "Go to questions");
             btnq.Clicked += OnButtonQuestionsClicked;
-            win.Add(btnu, btnq, userLabel);
+            win.Add(btnu, btnq, btnex, userLabel);
             Rect frame = new Rect(4, 8, top.Frame.Width, 200);
             if (answerRepository.GetTotalPages() < 1) 
             {
@@ -400,6 +525,11 @@ namespace ConsoleApp
             Application.Run(dialog);
             if (dialog.deleted)
             {
+                if (questionRepository.GetById(answer.questionId).userId != current.id && current.isModerator == "no")
+                {
+                    MessageBox.Query("Error", "You have no access to this function", "Ok");
+                    return;
+                }
                 int pageNumber = Int32.Parse(Convert.ToString(page.Text));
                 answerRepository.DeleteById(answer.id);
                 int total = answerRepository.GetTotalPages();
@@ -423,6 +553,11 @@ namespace ConsoleApp
                 int pageNumber = answerId/10;
                 if(answerId % 10 != 0) pageNumber++;
                 answer = dialog.GetAnswer();
+                if(answer.text == "") 
+                {
+                    MessageBox.Query("Error", "Wrong input", "Ok");
+                    return;
+                }
                 answerRepository.Update(answerId, answer);
                 answersListView.SetSource(answerRepository.GetPage(pageNumber));
            }
@@ -434,6 +569,11 @@ namespace ConsoleApp
             if(!dialog.canceled)
             {
                 Answer answer = dialog.GetAnswer();
+                if(answer.text == "") 
+                {
+                    MessageBox.Query("Error", "Wrong input", "Ok");
+                    return;
+                }
                 answerRepository.Insert(answer);
                 answersListView.SetSource(answerRepository.GetPage(answerRepository.GetTotalPages()));
                 page.Text = Convert.ToString(answerRepository.GetTotalPages());
@@ -449,7 +589,14 @@ namespace ConsoleApp
             {
                 if (dialog.path != "")
                 {
-                    Import.Read(dialog.path + @"\questions.xml", dialog.path + @"\answers.xml", questionRepository, answerRepository);
+                    try
+                    {
+                        Import.Read(dialog.path + @"\questions.xml", dialog.path + @"\answers.xml", questionRepository, answerRepository);
+                    }
+                    catch
+                    {
+                        MessageBox.Query("Error", "Files don`t exist", "Ok");
+                    }
                 }
                 else
                 {
@@ -457,6 +604,25 @@ namespace ConsoleApp
                 }
             }
 
+        }
+        static void OnGetImage()
+        {
+            ImageDialog dialog = new ImageDialog();
+            Application.Run(dialog);
+            if(!dialog.canceled)
+            {
+                int pinned = questionRepository.GetImagePinned(dialog.start, dialog.end);
+                int all = questionRepository.GetImageAll("2020-01-01", "2022-01-01");
+                var plt = new ScottPlot.Plot(600, 400);
+                double[] xs = {-1, 0, 1};
+                double[] valuesA = {0, all, 0};
+                double[] valuesB = {0, all+pinned, 0};
+                plt.PlotBar(xs, valuesB, label: "Pinned");
+                plt.PlotBar(xs, valuesA, label: "All");
+                plt.Legend();
+                plt.Title("Answers stats");
+                plt.SaveFig("AnswersStats.png");
+            }
         }
         static void OnExportClicked()
         {
@@ -473,6 +639,32 @@ namespace ConsoleApp
                     MessageBox.Query("Error", "Path was not selected", "Ok");
                 }
             }
+        }
+        static void OnChangeUser()
+        {
+            Toplevel top = Application.Top;
+            top.RemoveAll();
+            Window win = new Window("Log In")
+            {
+                X = 0,
+                Y = 1,
+                Width = Dim.Fill(),
+                Height = Dim.Fill() - 1
+            };
+            Application.Top.Add(win);
+            Button btnv = new Button(20, 15, "Verify");
+            btnv.Clicked += OnButtonVerifyClicked;
+            win.Add(btnv);
+            Button btnr = new Button(2, 1, "Registration");
+            btnr.Clicked += OnButtonRegClicked;
+            win.Add(btnr);
+            Label lLabel = new Label(2, 4, "Login: ");
+            inputLogin = new TextField(20, 4, 40, "");
+            win.Add(lLabel, inputLogin);
+            Label pLabel = new Label(2, 6, "Password: ");
+            inputPass = new TextField(20, 6, 40, "");
+            win.Add(pLabel, inputPass);
+            Application.Run();
         }
         static void OnAbout()
         {
